+++ src/lib/components/utilities/browserrender/index.ts
@@ -1,0 +1,1 @@
+export {default as BrowserRender} from "./BrowserRender.svelte";
