+++ src/lib/components/utilities/intersectionrender/index.ts
@@ -1,0 +1,1 @@
+export {default as IntersectionRender} from "./IntersectionRender.svelte";
