--- src/lib/components/utilities/lazyrender/index.ts
@@ -1,1 +1,0 @@
-export {default as LazyRender} from "./LazyRender.svelte";
