--- src/lib/components/utilities/nobrowser/index.ts
@@ -1,1 +1,0 @@
-export {default as NoBrowser} from "./NoBrowser.svelte";
