--- src/lib/components/utilities/noserver/index.ts
@@ -1,1 +1,0 @@
-export {default as NoServer} from "./NoServer.svelte";
