+++ src/lib/components/utilities/serverrender/index.ts
@@ -1,0 +1,1 @@
+export {default as ServerRender} from "./ServerRender.svelte";
