+++ src/lib/index.ts
@@ -82,12 +82,12 @@ export * from "./components/typography/code";
 export * from "./components/typography/heading";
 export * from "./components/typography/text";
 
+export * from "./components/utilities/browserrender";
 export * from "./components/utilities/contextbackdrop";
 export * from "./components/utilities/contextbutton";
-export * from "./components/utilities/lazyrender";
-export * from "./components/utilities/nobrowser";
-export * from "./components/utilities/noserver";
+export * from "./components/utilities/intersectionrender";
 export * from "./components/utilities/portal";
+export * from "./components/utilities/serverrender";
 export * from "./components/utilities/transition";
 
 export * from "./actions/click_outside";
@@ -95,7 +95,6 @@ export * from "./actions/keybind";
 export * from "./actions/intersection_observer";
 export * from "./actions/mutation_observer";
 
-export * from "./stores/attribute";
 export * from "./stores/component";
 export * from "./stores/darkmode";
 export * from "./stores/formstate";
